+++ jest.config.functional.js
@@ -5,5 +5,6 @@ module.exports = {
   preset: mainJestConfig.preset,
   roots: ['src/functional_tests'],
   testEnvironment: mainJestConfig.testEnvironment,
+  setupFiles: ['./src/functional_tests/jest.setup.ts'],
   setupFilesAfterEnv: mainJestConfig.setupFilesAfterEnv
 };
